Started work on adding notes at specific timestamps
