+++ client/src/components/NewNoteForm.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import styles from '../css-modules/EditFlowPage.module.css';
+
+const NewNoteForm = () => {
+    return (
+        <form>
+            <textarea className={styles.noteForm + ' form-control'} />
+        </form>
+    )
+}
+
+export default NewNoteForm;
