+++ client/src/components/NoteButton.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { MDBIcon } from 'mdbreact';
+
+const NoteButton = () => {
+    return (
+        <button className='btn btn-light'><MDBIcon icon='comment' /></button>
+    )
+}
+
+export default NoteButton;
